+++ manager.go
@@ -24,18 +24,28 @@ func (m *Manager) StartServer(server Server, lis net.Listener) func(ch chan<- er
 	return func(ch chan<- error) {
 		level.Info(m.Logger).Log(
 			"msg", "Starting server",
-			"addr", lis.Addr().String(),
+			"addr", lis.Addr(),
 			"server", getServerName(server),
 		)
+
 		ch <- server.Serve(lis)
 	}
 }
 
-// ListenAndStartServer creates a server starter function which listens on a port and can be called as a goroutine
-func (m *Manager) ListenAndStartServer(server Server, network string, addr string) func(ch chan<- error) {
-	lis, err := net.Listen(network, addr)
-	if err != nil {
-		panic(err)
+// ListenAndStartServer creates a server starter function which listens on a port and can be called as a goroutine.
+func (m *Manager) ListenAndStartServer(server Server, addr net.Addr) (func(ch chan<- error), error) {
+	var lis net.Listener
+	var err error
+
+	// Servers without address can still be managed via the Manager.
+	if addr != nil {
+		lis, err = net.Listen(addr.Network(), addr.String())
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		addr = NewAddr("none", "none")
+		lis = newVirtualListener(addr)
 	}
 
 	level.Info(m.Logger).Log(
@@ -44,10 +54,10 @@ func (m *Manager) ListenAndStartServer(server Server, network string, addr strin
 		"server", getServerName(server),
 	)
 
-	return m.StartServer(server, lis)
+	return m.StartServer(server, lis), nil
 }
 
-// StopServer creates a server stopper function which can be called as a goroutine
+// StopServer creates a server stopper function which can be called as a goroutine.
 func (m *Manager) StopServer(server Server, wg *sync.WaitGroup) func(ctx context.Context) error {
 	wg.Add(1)